ESHttpClient: Move HTTPS related code into ESHttpsClient class
