+++ modules/java-modules/elastic-v2/src/main/java/org/syslog_ng/elasticsearch_v2/client/ESClientFactory.java
@@ -29,6 +29,7 @@ import org.syslog_ng.elasticsearch_v2.client.esnative.ESTransportClient;
 import org.syslog_ng.elasticsearch_v2.client.esnative.ESTransportShieldClient;
 import org.syslog_ng.elasticsearch_v2.client.esnative.ESTransportSearchGuardClient;
 import org.syslog_ng.elasticsearch_v2.client.http.ESHttpClient;
+import org.syslog_ng.elasticsearch_v2.client.http.ESHttpsClient;
 
 public class ESClientFactory {
 	public static ESClient getESClient(ElasticSearchOptions options) throws UnknownESClientModeException {
@@ -46,7 +47,7 @@ public class ESClientFactory {
 			return new ESHttpClient(options);
 		}
 		else if (client_type.equals(ElasticSearchOptions.CLIENT_MODE_HTTPS)) {
-			return new ESHttpClient(options);
+			return new ESHttpsClient(options);
 		}
 		else if (client_type.equals(ElasticSearchOptions.CLIENT_MODE_SEARCHGUARD)) {
 			return new ESTransportSearchGuardClient(options);
